filter ghost snippets in export

+++ src/Sulu/Bundle/SnippetBundle/Command/SnippetExportCommand.php
@@ -15,7 +15,9 @@ use Sulu\Component\Snippet\Export\SnippetExportInterface;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 class SnippetExportCommand extends Command
 {
@@ -38,13 +40,46 @@ class SnippetExportCommand extends Command
         $this->setDescription('Export snippet translations from given language.');
         $this->addArgument('target', InputArgument::REQUIRED, 'Target for export (e.g. export_de.xliff)');
         $this->addArgument('locale', InputArgument::REQUIRED, 'Locale to export (e.g. de, en)');
+        $this->addOption('format', 'f', InputOption::VALUE_REQUIRED, '', '1.2.xliff');
     }
 
     public function execute(InputInterface $input, OutputInterface $output)
     {
         $target = $input->getArgument('target');
+        if (false === \strpos($target, '/')) {
+            $target = \getcwd() . '/' . $target;
+        }
         $locale = $input->getArgument('locale');
-        $file = $this->snippetExporter->export($locale, $output, '1.2.xliff');
+        $format = $input->getOption('format');
+
+        $output->writeln([
+            '<info>Language Export</info>',
+            '<info>===============</info>',
+            '',
+            '<info>Options</info>',
+            'Target: ' . $target,
+            'Locale: ' . $locale,
+            'Format: ' . $format,
+            '---------------',
+            '',
+        ]);
+
+        $helper = $this->getHelper('question');
+        $question = new ConfirmationQuestion('<question>Continue with this options?(y/n)</question> ', false);
+
+        if (!$helper->ask($input, $output, $question)) {
+            $output->writeln('<error>Abort!</error>');
+
+            return 0;
+        }
+
+        $output->writeln('<info>Continue!</info>');
+
+        $file = $this->snippetExporter->export(
+            $locale,
+            $output,
+            $format
+        );
 
         \file_put_contents($target, $file);
